`Foreign.JavaScript.EventLoop` now uses STM to implement an event loop that correctly handles concurrent calls to `runEval` and `callEval`.
Messages from the client are read and put into different queues.
Calls are put into a queue and processed in order.
Events are read from a queue and processed in order.

Events that arrive while a call is made or an event is being handled are labelled `Inconsistent`, because they reflect a state of the client that is no longer consistent with the server. This information is currently not used, but we may want to make use of it in the API.