`Foreign.RemotePtr.unprotectedGetCoupon` now lives in the `IO` monad.
This fixes a bug where a `JSObject` that is only referenced by `ffi` calls will be garbage collected, because only its `Coupon` is kept alive. By putting the internal `toCode` function in the `IO` monad, we ensure that the `RemotePtr` is kept alive, because we have to read from it whenever we want to make the FFI call.