Reenable garbage collection for ‘Element’s
The children and the event handlers associated to an ‘Element’
are now managed by two separate ‘RemotePtr’.

Previously, both the children of an ‘Element’ and the event handlers
associated to this element where “attached” to it
by using ‘Foreign.addReachable’. Clearing all child elements
had the nasty side effect of removing the event handlers as well.

Now, we can use ‘Foreign.clearReachable’ on the ‘elChildren’
pointer to sever the connection to the child elements,
allowing them to be garbage collected if there is no reference to them
on the server side anymore.

Fix #109, #113