Establish and implement synchronous semantics for `disconnect` event #133
This makes registering the disconnect handler much
more robust. On the other hand, it takes away the
possibility to interrupt a long running computation
by closing the browser window.

Also use the `commOpen` TVar to implement the `commClose`
function, instead of using `throwTo`, because
it is not clear what happens when we do that twice.