Improve printing of pure exceptions #145
The API may be called in ways where pure values
contain exceptions, e.g.

	element div # set UI.text (error “oops”)

This call will be translated into a message for
the JavaScript side. Sending this message will
fail, because it contains a _|_ value.
We now try to force this exception in the
translation phase, because this will result in a
better error message.